Move `resolveFieldRestrictions` function to its own file
